refactor : Player 의존성 제거
