+++ src/main/java/mafia/mafiatogether/job/domain/jobtype/Doctor.java
@@ -1,8 +1,6 @@
 package mafia.mafiatogether.job.domain.jobtype;
 
 import java.util.List;
-import java.util.Map;
-import mafia.mafiatogether.game.domain.Player;
 import mafia.mafiatogether.job.domain.JobTarget;
 import mafia.mafiatogether.job.domain.PlayerJob;
 
@@ -15,11 +13,6 @@ public class Doctor implements Job {
             final String targetName
     ) {
         return targetName;
-    }
-
-    public static void executeSkill(Map<JobType, Player> targets) {
-        final Player target = targets.getOrDefault(JobType.DOCTOR, Player.NONE);
-        target.heal();
     }
 
     @Override
